!!! TASK: Move `SearchTerm` to own namespace like other filters
